# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Notes API",
      "version": "1.0.0",
      "description": "API for managing notes and users"
    },
    "servers": [
      {
        "url": "http://localhost:8000",
        "description": "Local server"
      }
    ],
    "paths": {
      "/create-account": {
        "post": {
          "summary": "Create a new account",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fullname": { "type": "string" },
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["fullname", "email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful account creation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "boolean" },
                      "user": {
                        "type": "object",
                        "properties": {
                          "fullName": { "type": "string" },
                          "email": { "type": "string" },
                          "password": { "type": "string" },
                          "createdOn": { "type": "string", "format": "date-time" }
                        }
                      },
                      "accessToken": { "type": "string" },
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad requests",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "boolean" },
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "User login",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  },
                  "required": ["email", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful login",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "boolean" },
                      "message": { "type": "string" },
                      "email": { "type": "string" },
                      "password": { "type": "string" },
                      "accessToken": { "type": "string" }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad requests",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "boolean" },
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/get-user": {
        "get": {
          "summary": "Get user information",
          "security": [
            { "bearerAuth": [] }
          ],
          "responses": {
            "200": {
              "description": "Successful information retrieval",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "object",
                        "properties": {
                          "fullName": { "type": "string" },
                          "email": { "type": "string" },
                          "_id": { "type": "string" },
                          "createdOn": { "type": "string", "format": "date-time" }
                        }
                      },
                      "message": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
  